Refactor and integrate kmeans and nn
This merges Aiara's kmeans implementation with Maarten's nn algorithm.

Code for both algorithms has been refactored and modified for fixes.

Querying global (server) status seems to still be broken, as in I'm
unable to get that request (GET /status) through the data-app. So I have
hackily modified '/model' for this purpose.

This is still obviously very much quick proof-of-concept.